refactor to make manage vol opps a parent of both staff area and event edit
